backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
scikit-learn
joblib
weasyprint
ortools
